# swagger/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "File system Service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://{host}:{port}",
            "variables": {
                "host": {
                    "default": "localhost"
                },
                "port": {
                    "default": "3000",
                    "description": "Port where is exposed service."
                }
            }
        }
    ],
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "/auth/get-token",
                        "scopes": {
                            "readwrite": "lecture et ecriture",
                            "readonly": "lecture seule"
                        }
                    }
                }
            }
        },
        "requestBodies": {
            "FileSystemItem": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/FileSystemItem"
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "HttpError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "integer"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "FileSystemItemWithoutExtension": {
                "type": "object",
                "required": [
                    "type",
                    "path",
                    "name"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "file",
                            "directory"
                        ]
                    },
                    "path": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "symlink": {
                        "type": "string"
                    }
                }
            },
            "FileSystemItemExtension": {
                "type": "object",
                "properties": {
                    "extension": {
                        "type": "string",
                        "enum": [
                            "",
                            ".txt",
                            ".pdf",
                            ".json",
                            ".xml"
                        ]
                    }
                }
            },
            "FileSystemItem": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/FileSystemItemWithoutExtension"
                    },
                    {
                        "$ref": "#/components/schemas/FileSystemItemExtension"
                    }
                ]
            },
            "ResponseStatus": {
                "type": "object",
                "properties": {
                    "status": {
                        "enum": [
                            "success",
                            "error"
                        ]
                    }
                }
            },
...
        }
    },
    "security": [
        {
            "oauth2": [
                "readonly",
                "readwrite"
            ]
        }
    ],
...
    "paths": {
...
        "/file/{path}": {
            "parameters": [
                {
                    "name": "path",
                    "description": "Path of file to rename.",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "path",
                        "default": "home/nchoquet"
                    },
                    "style": "simple",
                    "example": "home/nchoquet"
                }
            ],
            "get": {
                "summary": "Get selected file content",
                "description": "Get selected file content",
                "security": [
                    {
                        "oauth2": [
                            "readonly",
                            "readwrite"
                        ]
                    }
                ],
                "tags": [
                    "Files"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            },
                            "text/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "Rename selected file",
                "description": "Rename selected file",
                "security": [
                    {
                        "oauth2": [
                            "readonly",
                            "readwrite"
                        ]
                    }
                ],
                "tags": [
                    "Files"
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/FileSystemItem"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileSystemItem"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update file content",
                "summary": "Update file content",
                "security": [
                    {
                        "oauth2": [
                            "readonly",
                            "readwrite"
                        ]
                    }
                ],
                "tags": [
                    "Files"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "type": "array"
                                    }
                                ]
                            }
                        },
                        "text/plain": {
                            "schema": {
                                "type": "string"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "type": "string"
                            }
                        },
                        "application/pdf": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseStatus"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete selected file",
                "summary": "Delete selected file",
                "security": [
                    {
                        "oauth2": [
                            "readonly",
                            "readwrite"
                        ]
                    }
                ],
                "tags": [
                    "Files"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseStatus"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpError"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}